rework inheiritance of Win32Registry to support unit test mocking
